- SQL IDE: spatial view: fixed a number of warnings. Corrected copyright information in some files. - Win: wbprivate.be: adjusted project settings for gdal. Created new filters for spatial files in project structure. - Win: updated PrepareOutputDir.cmd to copy gdal dll and helper apps. - Win: updated msi source to include gdal and their tools (ogrinfo and ogr2ogr for now).
[TESTS]
  - Ran a simple spatial query to show the drawing surface with a point on it.
  - Tried running an shp import, which failed in Python. Need to investigate.

+++ backend/wbprivate/sqlide/spatial_data_view.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2013, 2014, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2014, Oracle and/or its affiliates. All rights reserved.
  *
  * This program is free software; you can redistribute it and/or
  * modify it under the terms of the GNU General Public License as
@@ -64,7 +64,7 @@ public:
         std::string geom_data; // data in MySQL internal binary geometry format.. this is neither WKT nor WKB
         // but the internal format seems to be 4 bytes of SRID followed by WKB data
         if (rs->get_raw_field(row, _geom_column, geom_data) && !geom_data.empty())
-          add_feature(row, geom_data, false);
+          add_feature((int)row, geom_data, false);
       }
     }
   }
